pyproject.toml
[project]
name = "resnet-cifar10"
version = "0.1.0"
description = "Paper-faithful ResNet-18 on CIFAR-10 with residual ablation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.9.0",
    "numpy>=2.0.0",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
